added good form for merging
To merge the local version with the repo version just 'git pull' right before you 'git push'.  
If we have both edited the same line git adds a tag to the local file indicating the conficting lines as well as the two versions of the conflicting lines.
We must then go through the file again and resolve all of these tags.
After conflict resolution 'git pull' will exectute as "Already up-to-date" and 'git push' will succeed.